+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- `ExplainCommand`
+	- explains cron expression syntax
 - `ListCommand`
 	- adds `--explain` option to explain whole expression